# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentDirection": {
        "enum": [
          "flooding",
          "ebbing"
        ],
        "title": "CurrentDirection",
        "type": "string"
      },
      "CurrentInfo": {
        "additionalProperties": false,
        "description": "Current prediction/observation information for API responses.",
        "properties": {
          "direction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentDirection"
              },
              {
                "type": "null"
              }
            ],
            "description": "Direction of current (Enum: CurrentDirection, null if non-tidal)"
          },
          "magnitude": {
            "description": "Current strength in knots",
            "title": "Magnitude",
            "type": "number"
          },
          "magnitude_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Relative current strength (0.0-1.0, null if non-tidal), normalized against the nearest local flood/ebb peak in the available prediction curve.",
            "title": "Magnitude Pct"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentPhase"
              },
              {
                "type": "null"
              }
            ],
            "description": "Compact current phase for displays. Tidal prediction values are flood, ebb, slack_before_flood, slack_before_ebb, or slack. Slack phases are assigned when absolute current magnitude is below 0.2 knots; slack_before_* indicates the next non-slack direction predicted by the curve."
          },
          "source_type": {
            "$ref": "#/components/schemas/DataSourceType",
            "description": "Indicates if data is prediction or observation (Enum: DataSourceType)"
          },
          "state_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Display-ready current state phrase, such as 'strong ebb and building' or 'slack before flood' (null if non-tidal).",
            "title": "State Description"
          },
          "strength": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentStrength"
              },
              {
                "type": "null"
              }
            ],
            "description": "Cycle-relative current strength for non-slack tidal predictions (light, moderate, or strong). Null for slack or non-tidal currents."
          },
          "timestamp": {
            "description": "ISO 8601 formatted timestamp of the prediction (in location's local timezone)",
            "title": "Timestamp",
            "type": "string"
          },
          "trend": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentTrend"
              },
              {
                "type": "null"
              }
            ],
            "description": "Current trend for non-slack tidal predictions (building, easing, or steady). Null for slack or non-tidal currents."
          }
        },
        "required": [
          "timestamp",
          "magnitude",
          "source_type"
        ],
        "title": "CurrentInfo",
        "type": "object"
      },
      "CurrentPhase": {
        "description": "Compact current phase for API consumers and displays.",
        "enum": [
          "flood",
          "ebb",
          "slack_before_flood",
          "slack_before_ebb",
          "slack"
        ],
        "title": "CurrentPhase",
        "type": "string"
      },
      "CurrentStrength": {
        "description": "Cycle-relative current strength for displays.",
        "enum": [
          "light",
          "moderate",
          "strong"
        ],
        "title": "CurrentStrength",
        "type": "string"
      },
      "CurrentTrend": {
        "description": "Whether the current is building or easing.",
        "enum": [
          "building",
          "easing",
          "steady"
        ],
        "title": "CurrentTrend",
        "type": "string"
      },
      "CurrentsResponse": {
        "additionalProperties": false,
        "description": "Complete response for currents API endpoint.",
        "properties": {
          "current": {
            "$ref": "#/components/schemas/CurrentInfo"
          },
          "current_chart_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Filename of the current chart image (only for locations with chart assets)",
            "title": "Current Chart Filename"
          },
          "legacy_chart": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegacyChartInfo"
              },
              {
                "type": "null"
              }
            ],
            "description": "Legacy chart info (only for locations with chart assets)"
          },
          "location": {
            "$ref": "#/components/schemas/LocationInfo"
          },
          "navigation": {
            "additionalProperties": true,
            "description": "Navigation parameters for time shifting",
            "title": "Navigation",
            "type": "object"
          },
          "timestamp": {
            "description": "ISO 8601 formatted timestamp of the prediction (in location's local timezone)",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "location",
          "timestamp",
          "current",
          "navigation"
        ],
        "title": "CurrentsResponse",
        "type": "object"
      },
...
      "DataSourceType": {
        "description": "Indicates whether data is based on a prediction or an observation.",
        "enum": [
          "prediction",
          "observation"
        ],
        "title": "DataSourceType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LegacyChartInfo": {
        "additionalProperties": false,
        "description": "Information about legacy tide charts for API responses.",
        "properties": {
          "chart_filename": {
            "description": "Filename of the legacy chart",
            "title": "Chart Filename",
            "type": "string"
          },
          "hours_since_last_tide": {
            "description": "Hours since the last tide event",
            "title": "Hours Since Last Tide",
            "type": "number"
          },
          "last_tide_type": {
            "$ref": "#/components/schemas/TideCategory",
            "description": "Type of last tide (Enum: TideCategory)"
          },
          "map_title": {
            "description": "Title for the legacy map",
            "title": "Map Title",
            "type": "string"
          }
        },
        "required": [
          "hours_since_last_tide",
          "last_tide_type",
          "chart_filename",
          "map_title"
        ],
        "title": "LegacyChartInfo",
        "type": "object"
      },
...
      "LocationInfo": {
        "additionalProperties": false,
        "description": "Location information for API responses.",
        "properties": {
          "code": {
            "description": "Location code (e.g., 'nyc')",
            "title": "Code",
            "type": "string"
          },
          "name": {
            "description": "Display name of the location",
            "title": "Name",
            "type": "string"
          },
          "swim_location": {
            "description": "Specific swimming location",
            "title": "Swim Location",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "swim_location"
        ],
        "title": "LocationInfo",
        "type": "object"
      },
...
      "TideCategory": {
        "enum": [
          "low",
          "high"
        ],
        "title": "TideCategory",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/{location}/currents": {
      "get": {
        "description": "API endpoint that returns current predictions for a specific location.\n\nArgs:\n    location: Location code (e.g., 'nyc')\n    shift: Time shift in minutes from current time (optional)\n\nReturns:\n    CurrentsResponse object with current prediction details\n\nRaises:\n    HTTPException: If the location is not configured or doesn't support currents",
        "operationId": "location_currents_api__location__currents_get",
        "parameters": [
          {
            "in": "path",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "shift",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Shift",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location Currents"
      }
    },
...
    "/{location}/currents": {
      "get": {
        "description": "Serve the water current visualization page for the specified location.\n\nArgs:\n    request: FastAPI request object\n    location: Location code (e.g., \"nyc\", \"san\")\n    shift: Time shift in minutes from current time\n\nReturns:\n    HTML response with current water conditions",
        "operationId": "location_currents__location__currents_get",
        "parameters": [
          {
            "in": "path",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "shift",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Shift",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location Currents"
      }
    },
    "/{location}/embed": {
      "get": {
        "description": "Serve the embed view for a specific location for embedding in other websites.\n\nArgs:\n    request: FastAPI request object\n    location: Location code (e.g., 'nyc')\n\nReturns:\n    HTML response with location-specific embed\n\nRaises:\n    HTTPException: If the location is not configured",
        "operationId": "location_embed__location__embed_get",
        "parameters": [
          {
            "in": "path",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location Embed"
      }
    },
    "/{location}/widget": {
      "get": {
        "description": "Serve a widget view for a specific location.\n\nArgs:\n    request: FastAPI request object\n    location: Location code (e.g., 'nyc')\n\nReturns:\n    HTML response with location-specific widget\n\nRaises:\n    HTTPException: If the location is not configured",
        "operationId": "location_widget__location__widget_get",
        "parameters": [
          {
            "in": "path",
            "name": "location",
            "required": true,
            "schema": {
              "title": "Location",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location Widget"
      }
    }
  }
}